REDCOMSITE-325: Send information for alert_notifications plugin (low stock alert)
