--- plugins/redshop_alert/alert_notifications/language/en-GB/en-GB.plg_redshop_alert_alert_notifications.ini
@@ -1,6 +1,0 @@
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_TITLE="Custom Field Min Stock:"
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_MESSAGE="Product: <strong>%s</strong> (Product Number: %s )  has remaining %s in stock (%s = %s) </a>"
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_MESSAGE_DELETE="Warning Product:<strong>%s</strong>  deleted in stock (by user %s deleted).</a>"
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_TEMPLATE="Template Mail Min Stock"
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_MESSAGE_FROM_NAME="Administrator"
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_MESSAGE_SUBJECT="Min Stock"
