--- plugins/redshop_alert/alert_notifications/language/en-GB/en-GB.plg_redshop_alert_alert_notifications.sys.ini
@@ -1,2 +1,0 @@
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK="redSHOP: Alert Min Stock"
-PLG_REDSHOP_ALERT_ALERT_DESC_MIN_STOCK="Alert to administrator about user activities"
