--- plugins/redshop_alert/alert_notifications/language/lt-LT/lt-LT.plg_redshop_alert_alert_notifications.ini
@@ -1,6 +1,0 @@
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_TITLE="Custom Field Min Stock:"
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_MESSAGE="Productas: <strong>%s</strong> (Product Number: %s )  likutis %s (%s = %s) </a>"
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_MESSAGE_DELETE="Warning Product:<strong>%s</strong>  deleted in stock (by user %s deleted).</a>"
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_TEMPLATE="Template Mail Min Stock"
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_MESSAGE_FROM_NAME="Bitė MRKT Admin"
-PLG_REDSHOP_ALERT_ALERT_MIN_STOCK_MESSAGE_SUBJECT="Dėmesio! Mažas atsargų likutis"
