+++ plugins/redshop_alert/low_stock_alert/language/en-GB/en-GB.plg_redshop_alert_low_stock_alert.ini
@@ -1,0 +1,6 @@
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_TITLE="Custom Field Low Stock:"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_MESSAGE="Product: <strong>%s</strong> (Product Number: %s )  has remaining %s in stock (%s = %s) </a>"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_MESSAGE_DELETE="Warning Product:<strong>%s</strong>  deleted in stock (by user %s deleted).</a>"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_TEMPLATE="Template Mail Low Stock"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_MESSAGE_FROM_NAME="Administrator"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_MESSAGE_SUBJECT="Low Stock"
