+++ plugins/redshop_alert/low_stock_alert/language/lt-LT/lt-LT.plg_redshop_alert_low_stock_alert.ini
@@ -1,0 +1,6 @@
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_TITLE="Custom Field Min Stock:"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_MESSAGE="Productas: <strong>%s</strong> (Product Number: %s )  likutis %s (%s = %s) </a>"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_MESSAGE_DELETE="Warning Product:<strong>%s</strong>  deleted in stock (by user %s deleted).</a>"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_TEMPLATE="Template Mail Min Stock"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_MESSAGE_FROM_NAME="Bitė MRKT Admin"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_MESSAGE_SUBJECT="Dėmesio! Mažas atsargų likutis"
