+++ plugins/redshop_alert/low_stock_alert/language/lt-LT/lt-LT.plg_redshop_alert_low_stock_alert.sys.ini
@@ -1,0 +1,4 @@
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT="redSHOP: Low Stock Alert"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_DESC="Alert to administrator about user activities"
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_TITLE = 'Title : Low Stock Alert'
+PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_TEMPLATE = 'Tag : Low Stock Alert'
