+++ plugins/redshop_alert/low_stock_alert/low_stock_alert.xml
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="utf-8"?>
+<extension version="3.5.1" type="plugin" group="redshop_alert" method="upgrade">
+    <name>PLG_REDSHOP_ALERT_LOW_STOCK_ALERT</name>
+    <version>2.0.0</version>
+    <redshop>2.1.2.2</redshop>
+    <creationDate>August 2019</creationDate>
+    <author>redCOMPONENT.com</author>
+    <authorEmail>email@redcomponent.com</authorEmail>
+    <authorUrl>www.redcomponent.com</authorUrl>
+    <copyright>(C) 2019 redCOMPONENT.com</copyright>
+    <license>GNU/GPL</license>
+    <description>PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_DESC</description>
+    <languages>
+        <language tag="en-GB">language/en-GB/en-GB.plg_redshop_alert_low_stock_alert.ini</language>
+        <language tag="en-GB">language/en-GB/en-GB.plg_redshop_alert_low_stock_alert.sys.ini</language>
+        <language tag="lt-LT">language/lt-LT/lt-LT.plg_redshop_alert_low_stock_alert.ini</language>
+        <language tag="lt-LT">language/lt-LT/lt-LT.plg_redshop_alert_low_stock_alert.sys.ini</language>
+    </languages>
+    <files>
+        <filename plugin="low_stock_alert">low_stock_alert.php</filename>
+        <filename>index.html</filename>
+        <folder>sql</folder>
+    </files>
+    <install>
+        <sql>
+            <file driver="mysql" charset="utf8">sql/install.sql</file>
+        </sql>
+    </install>
+    <config>
+        <fields name="params">
+            <fieldset name="basic">
+                <field
+                        name="id_min_stock"
+                        type="sql"
+                        default="10"
+                        label="PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_TITLE"
+                        query="SELECT  name , id  FROM #__redshop_fields where section=1 and published=1 "
+                        key_field="id"
+                        value_field="name"
+                />
+                <field
+                        name="id_alert_min_stock_template"
+                        type="sql"
+                        default="10"
+                        label="PLG_REDSHOP_ALERT_LOW_STOCK_ALERT_TEMPLATE"
+                        query="SELECT  `name` , `id`  FROM `#__redshop_template`  where  published=1 "
+                        key_field="id"
+                        value_field="name"
+                />
+            </fieldset>
+        </fields>
+    </config>
+</extension>
