+++ plugins/redshop_alert/low_stock_alert/sql/install.sql
@@ -1,0 +1,5 @@
+INSERT INTO `#__redshop_fields` (`id`, `title`, `name`, `type`, `desc`, `class`, `section`, `groupId`, `maxlength`, `cols`, `rows`, `size`, `show_in_front`, `required`, `published`, `publish_up`, `publish_down`, `display_in_product`, `ordering`, `display_in_checkout`, `checked_out`, `checked_out_time`, `created_date`, `created_by`, `modified_date`, `modified_by`)
+VALUES (NULL, 'Low Stock Alert', 'rs_low_stock_alert', '1', '', '', '1', NULL, '1000', '0', '0', '100', '1', '0', '1', '2019-08-28 09:35:47', '2019-08-28 09:35:47', '1', '7', '1', '0', '2019-08-28 09:35:47', '2019-08-28 09:35:47', '249', '2019-08-28 09:35:47', '249');
+
+INSERT INTO `#__redshop_template` (`id`, `name`, `section`, `file_name`, `order_status`, `payment_methods`, `published`, `shipping_methods`, `checked_out`, `checked_out_time`, `created_by`, `created_time`, `modified_by`, `modified_date`)
+VALUES (NULL, 'low_stock_template', 'product', '1_low_stock_template', '', '', '1', '', '249', '2019-08-28 09:45:04', '249', '2019-08-28 09:45:04', '249', '2019-08-28 09:45:04');
