+++ tests/_pages/Joomla3/Administrator/ConfigurationPage.php
@@ -520,88 +520,4 @@ class ConfigurationPage extends AdminJ3Page
 	 * @since 2.1.2
 	 */
 	public static $twoWayRelatedNo = '//label[@id="twoway_related_product_0-lbl"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $ordersTab = '//a[@href="#ordertab"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $resetOderId = '//a[@title="Order ID Reset"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $sendOrderEmail = '//div[@id="s2id_order_mail_after"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $inputOderEmail = '//input[@id="s2id_autogen44_search"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $enableInvoiceEmailYes = '//label[@id="invoice_mail_enable_1-lbl"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $enableInvoiceEmailNo = '//label[@id="invoice_mail_enable_0-lbl"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $sendMailToCustomerInOrderYes = '//label[@id="send_mail_to_customer_1-lbl"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $sendMailToCustomerInOrderNo = '//label[@id="send_mail_to_customer_0-lbl"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $afterPayment = '//ul[@id="select2-results-44"]/li[3]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $noneButton = '//label[@id="invoice_mail_send_option_0-lbl"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $administratorButton = '//label[@id="invoice_mail_send_option_1-lbl"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $customerButton = '//label[@id="invoice_mail_send_option_2-lbl"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $bothButton = '//label[@id="invoice_mail_send_option_3-lbl"]';
-
-	/**
-	 * @var string
-	 * @since 2.1.3
-	 */
-	public static $messagePopup = 'Successfully reset order id';
 }
