+++ tests/acceptance/administrator/g07/Configuration/ConfigurationCest.php
@@ -1,89 +1,12 @@
 <?php
-/**
- * @package     redSHOP
- * @subpackage  Cest
- * @copyright   Copyright (C) 2008 - 2019 redCOMPONENT.com. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-use Configuration\ConfigurationSteps;
 
 /**
- * Class ConfigurationCest
- * @since 2.1.3
+ *
+ * Configuration function
+ *
  */
 class ConfigurationCest
 {
-	/**
-	 * @var array
-	 * @since 2.1.3
-	 */
-	protected $configurationOder1;
-
-	/**
-	 * @var array
-	 * @since 2.1.3
-	 */
-	protected $configurationOder2;
-
-	/**
-	 * @var array
-	 * @since 2.1.3
-	 */
-	protected $configurationOder3;
-
-	/**
-	 * @var array
-	 * @since 2.1.3
-	 */
-	protected $configurationOder4;
-
-	/**
-	 * @var array
-	 * @since 2.1.3
-	 */
-	protected $configurationOder5;
-
-	/**
-	 * @var array
-	 * @since 2.1.3
-	 */
-	protected $configurationOder6;
-
-	/**
-	 * @var array
-	 * @since 2.1.3
-	 */
-	protected $configurationOder7;
-
-	/**
-	 * @var array
-	 * @since 2.1.3
-	 */
-	protected $configurationOder8;
-
-	/**
-	 * @var array
-	 * @since
-	 */
-	protected $configurationOder9;
-
-	/**
-	 * @var array
-	 * @since 2.1.3
-	 */
-	protected $configurationOder10;
-
-	/**
-	 * @var array
-	 * @since 2.1.3
-	 */
-	protected $configurationOder11;
-
-	/**
-	 * ConfigurationCest constructor.
-	 * @since 2.1.3
-	 */
 	public function __construct()
 	{
 		$this->faker = Faker\Factory::create();
@@ -113,108 +36,20 @@ class ConfigurationCest
 		$this->quantityInCart   = 0;
 		$this->minimunOrder     = 0;
 		$this->enableQuation    = 'no';
-
-		//Configuration Oder Reset Id Oder
-		$this->configurationOder1 =
-			[
-				'resetIdOder'                   => 'Reset Id Oder',
-			];
-
-		//Configuration Oder Oder Email After Payment
-		$this->configurationOder2 =
-			[
-				'sendOderEmail'                 => 'Send Oder Email',
-				'afterPayment'                  => 'After Payment',
-			];
-
-		//Configuration Oder Send Oder email After Payment, But Send Before To Administrator
-		$this->configurationOder3 =
-			[
-				'sendOderEmail'                 => 'Send Oder Email',
-				'afterPayment2'                 => 'After Payment, but send before to administrator',
-			];
-
-
-		//Configuration Oder Send Oder Email Before Payment
-		$this->configurationOder4 =
-			[
-				'sendOderEmail'                 => 'Send Oder Email',
-				'beforePayment'                 => 'Before Payment',
-			];
-
-		//Configuration Oder Enable In Voice Email Yes, None
-		$this->configurationOder5 =
-			[
-				'enableInVoiceEmail'            => 'Enable In Voice Email',
-				'Yes'                           => 'Yes',
-				'None'                          => 'None',
-			];
-
-		//Configuration Oder Enable In Voice Email Yes, Administrator
-		$this->configurationOder6 =
-			[
-				'enableInVoiceEmail'            => 'Enable In Voice Email',
-				'Yes'                           => 'Yes',
-				'Administrator'                 => 'Administrator',
-			];
-
-		//Configuration Oder Enable In Voice Email Yes, Customer
-		$this->configurationOder7 =
-			[
-				'enableInVoiceEmail'            => 'Enable In Voice Email',
-				'Yes'                           => 'Yes',
-				'Customer'                      => 'Customer',
-			];
-
-		//Configuration Oder Enable In Voice Email Yes, Both
-		$this->configurationOder8 =
-			[
-				'enableInVoiceEmail'            => 'Enable In Voice Email',
-				'Yes'                           => 'Yes',
-				'Both'                          => 'Both',
-			];
-
-		//Configuration OderEnable In Voice Email, No
-		$this->configurationOder9 =
-			[
-				'enableInVoiceEmail'            => 'Enable In Voice Email',
-				'No'                            => 'No',
-			];
-
-		//Configuration Oder Send Mail To Customer In Oder, Yes
-		$this->configurationOder10 =
-			[
-				'sendMailToCustomerInOder'      => 'Send mail to customer in oder',
-				'Yes'                           => 'Yes',
-			];
-
-		//Configuration Oder Send Mail To Customer In oder No
-		$this->configurationOder11 =
-			[
-				'sendMailToCustomerInOder'      => 'Send mail to customer in oder',
-				'No'                            => 'No',
-			];
 	}
 
 	/**
 	 * @param AcceptanceTester $I
-	 * @throws Exception
-	 * @since 2.1.3
 	 */
 	public function _before(AcceptanceTester $I)
 	{
 		$I->doAdministratorLogin();
 	}
 
-	/**
-	 * @param AcceptanceTester $I
-	 * @param $scenario
-	 * @throws Exception
-	 * @since 2.1.3
-	 */
-	public function allCaseAtConfigurations(ConfigurationSteps $I)
+	public function allCaseAtConfigurations(AcceptanceTester $I, $scenario)
 	{
 		$I->wantTo('Test enable Stockroom in Administrator');
+		$I = new Configuration\ConfigurationSteps($scenario);
 		$I->featureUsedStockRoom();
 
 		$I->wantTo('Test off Stockroom in Administrator');
@@ -243,38 +78,5 @@ class ConfigurationCest
 
 		$I->wantTo('setup VAT at admin');
 		$I->cartSetting($this->addcart, $this->allowPreOrder, $this->enableQuation, $this->cartTimeOut, $this->enabldAjax, $this->defaultCart, $this->buttonCartLead, $this->onePage, $this->showShippingCart, $this->attributeImage, $this->quantityChange, $this->quantityInCart, $this->minimunOrder);
-
-		$I->wantTo('Test Configuration Oder - Reset Id Oder');
-		$I->ConfigurationOder($this->configurationOder1);
-
-		$I->wantTo('Test Configuration Oder - Oder Email After Payment');
-		$I->ConfigurationOder($this->configurationOder2);
-
-		$I->wantTo('Test Configuration Oder - Send Oder email After Payment, But Send Before To Administrator');
-		$I->ConfigurationOder($this->configurationOder3);
-
-		$I->wantTo('Test Configuration Oder - Send Oder Email Before Payment');
-		$I->ConfigurationOder($this->configurationOder4);
-
-		$I->wantTo('Test Configuration Oder - Enable In Voice Email Yes None');
-		$I->ConfigurationOder($this->configurationOder5);
-
-		$I->wantTo('Test Configuration Oder - Enable In Voice Email Yes Administrator');
-		$I->ConfigurationOder($this->configurationOder6);
-
-		$I->wantTo('Test Configuration Oder - Enable In Voice Email Yes Customer');
-		$I->ConfigurationOder($this->configurationOder7);
-
-		$I->wantTo('Test Configuration Oder - Enable In Voice Email Yes Both');
-		$I->ConfigurationOder($this->configurationOder8);
-
-		$I->wantTo('Test Configuration Oder - Enable In Voice Email No');
-		$I->ConfigurationOder($this->configurationOder9);
-
-		$I->wantTo('Test Configuration Oder - Send Mail To Customer In Oder Yes');
-		$I->ConfigurationOder($this->configurationOder10);
-
-		$I->wantTo('Test Configuration Oder - Send Mail To Customer in Oder No');
-		$I->ConfigurationOder($this->configurationOder11);
 	}
 }
